Improve move_mask and boolean ops for i64x2 i32x4 i16x8 i8x16 (#162)
* bug in scalar behavior of f32::fast_min/fast_max

* fixing max

* neon

* simd

* aarch

* aarch

* neon

* update wasmtime version

* fix wasm

* fix

* fix movemask

* syntax

* fixed fast_max

* revert yaml

* movemask and shr

* movemask only

* fix arm

* fix neon

* fix neon

* add neon to i64x2

* centralize move_mask ops in integer code

* update wasmtime version

* update wasmtime version

* add tracing

* more tracing for all

* added i64 tests for movemask

* improve tests

* fixed

* revert build changes

* remove float changes for now

* better neon version

* remove inefficient neon implemenation

+++ src/i8x32_.rs
@@ -294,7 +294,7 @@ impl i8x32 {
   pub fn move_mask(self) -> i32 {
     pick! {
       if #[cfg(target_feature="avx2")] {
-        move_mask_i8_m256i(self.avx) as i32
+        move_mask_i8_m256i(self.avx)
       } else {
         self.a.move_mask() | (self.b.move_mask() << 16)
       }
